automation/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
gspread==6.1.4
google-auth==2.35.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
